VGALIB cleanup and SDL event loop improvement
Cleaned up the constructor code for vga class, removed 1 char
allocations for buffer and initialized them to NULL.  Consolidated some
of the #ifdef code to make it cleaner, moved some of the initialization
around.
In the future it should be possible to draw sprites to the bg layer.
Revised the demo code to create a clock accurate event loop monitor to
process SDL events quickly enough.  The main loop wait is done in SDL
event handler, while the animate routine is called ~100hz.  Actual FPS
is determined by vsync.