Canvas optimizations and enhancements
Optimized the drawimage method to intelligently handle images that are a
subset of the destination image.  So if the src image is the same width
as the destination, and there is no x offset, it will just perform a
block copy instead of decomposing the image line by line.

Added another drawimage variant that accepts a subset of a src image, so
you can pan large images across the screen by simply adjusting the start
position in the src image.