Rework of event handling
The SDL WaitEvent(Timeout) will randomly add 10ms of delay to the event
loop, this is not functional in a loop that requires 60fps to keep up
with vsync.  I reimplemented the WaitEventTimeout in the vga class with
busy waiting and ms level delay precision.  This reliably drains events
from the SDL event queue and processes events quickly.  No noticable
increase in CPU usage observed.